backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1

# Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.1
pgvector==0.2.5

# Auth / Security
PyJWT[crypto]==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-multipart==0.0.9
httpx==0.27.0

# Redis / Job Queue
redis==5.0.4
rq==1.16.2

# OCR
pytesseract==0.3.10
Pillow==10.3.0
pdfplumber==0.11.0

# AI / Embeddings
openai==1.30.1

# Object Storage
boto3==1.34.107
cryptography==42.0.7

# Monitoring / Infra
sentry-sdk[fastapi]==2.1.1
slowapi==0.1.9

# SMS OTP (required when OTP_PROVIDER=twilio)
twilio==9.3.7

# Dev / Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
factory-boy==3.3.0
ruff==0.4.4
